+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Blue Triangle 2.19.1 Latest, Jan 5 2026
+### New Improvements
+- Added remote configuration flags to enable/disable following features:
+  - Grouping tap detection
+  - Network state tracking
+  - ANR tracking
+  - Crash tracking
+  - Memory warning
+  - Launch time tracking
+  - WebView stitching.
+- Changed default groupedViewSampleRate to 5%
+- Fixed excessive Launch time values issue. In cases where process was already created in background due to a work manager, push notification, BroadcastReceiver, etc., the Launch time was showing excessive values in hours and days even.
+
 # Blue Triangle 2.19.0 Latest, Dec 16 2025
 ### Improvements and Breaking Changes
 - Updated minimum supported Android version to API 21 (minSdkVersion = 21).